+++ R/split_phenotypes.R
@@ -21,7 +21,8 @@ utils::globalVariables(c(
 #' directory.
 #'
 #' The
-#' individual files must all have exactly the same `Sample Name` and `Cell ID`
+#' individual files must all have exactly the same
+#' `Sample Name` or `Annotation ID` and `Cell ID`
 #' columns. [split_phenotypes] is called to split the `Phenotype` columns.
 #'
 #' @param csd_files A list or vector of paths to cell seg data files.
@@ -49,14 +50,20 @@ consolidate_and_summarize_cell_seg_data = function(csd_files, output_dir,
 
   # Make some names, these will be for files and headers
   names = make_unique_names(csd_files)
+  field_col = NULL
 
   # Function to read a file, create a summary report, split phenotypes
   process_one_file <- function(name, path) {
     update_progress(detail=paste0('Reading "', name, '".'))
     d = phenoptr::read_cell_seg_data(path)
 
-    if (any(!c('Sample Name', 'Cell ID') %in% names(d)))
-      stop('Consolidation requires "Sample Name" and "Cell ID" columns in each data file.')
+    # Figure out the field column name from the first file
+    if (is.null(field_col))
+      field_col <<- phenoptr::field_column(d)
+
+    if (any(!c(field_col, 'Cell ID') %in% names(d)))
+      stop('Consolidation requires "', field_col,
+           '" and "Cell ID" columns in each data file.')
 
     update_progress(detail=paste0('Writing report for "', name, '".'))
     out_path = file.path(output_dir, paste0(name, '.html'))
@@ -72,15 +79,16 @@ consolidate_and_summarize_cell_seg_data = function(csd_files, output_dir,
   # Read subsequent files, report, split phenotypes, join with the first file.
   purrr::walk2(names[-1], csd_files[-1], function(name, path) {
     csd2 = process_one_file(name, path) %>%
-      dplyr::select(`Sample Name`, `Cell ID`, dplyr::starts_with('Phenotype '))
+      dplyr::select(!!rlang::sym(field_col), `Cell ID`,
+                    dplyr::starts_with('Phenotype '))
 
     if (nrow(csd2) != start_row_count)
       stop('Number of rows in data frames do not match.\n',
            nrow(csd2), ' != ', start_row_count, ' Failed at\n', path)
-    csd <<- dplyr::inner_join(csd, csd2, by=c('Sample Name', 'Cell ID'))
+    csd <<- dplyr::inner_join(csd, csd2, by=c(field_col, 'Cell ID'))
 
     if (nrow(csd) != start_row_count)
-      stop('Sample Names or Cell IDs do not match (rows dropped in join).\n',
+      stop(field_col, 's or Cell IDs do not match (rows dropped in join).\n',
            nrow(csd), ' != ', start_row_count, ' Failed at\n', path)
   })